requirements.txt
Django==1.6.11
GDAL==1.9.1
MultipartPostHandler==0.1.0
OWSLib==0.10.0
Paver==1.2.1
Pillow==3.0.0
Shapely==1.3.1
South==1.0.2
Unidecode==0.4.18
amqp==1.4.8
amqp==1.4.9
anyjson==0.3.3
awesome-slugify==1.6.2
beautifulsoup4==4.2.1
billiard==3.3.0.23
celery==3.1.17
Django==1.6.11
django-activity-stream==0.4.5beta1
django-appconf==0.5
django-auth-ldap==1.2.11
django-autocomplete-light==2.2.10
django-bootstrap3-datetimepicker==2.2.3
django-celery==3.1.16
django-downloadview==1.2
django-extensions==1.2.5
django-forms-bootstrap==3.0.1
django-friendly-tag-loader==1.1
django-geoexplorer==4.0.5
django-guardian==1.2.0
django-haystack==2.1.0
django-jsonfield==0.9.12
django-leaflet==0.13.7
django-modeltranslation==0.8
django-mptt==0.6.1
django-nose==1.2
django-pagination==1.0.7
django-polymorphic==0.5.6
django-taggit==0.12
django-tastypie==0.11.0
flake8==2.3.0
GDAL==1.11.2
geolinks==0.0.1
geonode-agon-ratings==0.3.4
geonode-announcements==1.0.5
geonode-arcrest==10.2
geonode-avatar==2.1.4
geonode-dialogos==0.4
geonode-notification==1.1.1
geonode-user-accounts==1.0.10
geonode-user-messages==0.1.2
gisdata==0.5.4
gsconfig==1.0.3
gsimporter==1.0.0
httplib2==0.10.3
kombu==3.0.37
lxml==3.4.0
mccabe==0.6.1
MultipartPostHandler==0.1.0
nose==1.3.1
olefile==0.44
OWSLib==0.10.0
Paver==1.2.1
pep8==1.6.2
Pillow==4.0.0
pinax-theme-bootstrap==3.0a11
pinax-theme-bootstrap-account==1.0b2
psycopg2==2.7.1
pycsw==1.10.3
pyelasticsearch==0.6.1
pyflakes==1.5.0
pyproj==1.9.5.1
python-cas==1.1.0
python-dateutil==2.6.0
python-ldap==2.4.38
python-mimeparse==1.6.0
pytz==2016.10
regex==2017.02.08
requests==2.13.0
Shapely==1.3.1
simplejson==3.10.0
six==1.10.0
South==1.0.2
transifex-client==0.10
Unidecode==0.04.20
wsgiref==0.1.2
